app/requirements.txt
fastapi
uvicorn
twilio
openai-whisper
requests
python-multipart
pydantic
python-dotenv
